Crate rusty_mos

The MOS operation system written in Rust. The origin OS is written in C and is for BUAA OS Course.

This version of MOS is independently completed by kai_Ker.

The C-Version of MOS is developed and continuously improved by the OS course team. I express my sincere gratitude to the teachers and *S.T.A.R.*s who have continuously contributed to the MOS and to out course.

Beyond the origin functions, this version of MOS also has a buddy system allocator and a multi-pages shared memory pool.

The MOS itself has the design concept of microkernel. The fork and the file system are done mainly in the user space. The MOS uses a page-memory management to organize the virtual memories.

Modules

arch

Things related to the architecture. Mipsel here is.

consts

Define constants used in the kernel.

kernel

kernel_tests

library

Kernel library. Only used in kernel mode.

memory

Memory management. Including the page-memory model and the TLB related handlers. A global allocator will be provided as well.

process

Envs manegement and the scheduler. Also support the elf_loader tools.

utils

Kernel independent utils like link list or bit map.

Macros

ENVX

Spawn the index of the given envid.

GEN_MASK

Gen a mask from the $h-bit (MSB) to $l-bit (LSB) with 1.

KADDR

Get the virtual address (in kernel segment) from the physical address

PADDR

Get the physical address of the virtual address in kernel segment

PDX

Get the Page Directory Offset from the virtual address

PPN

Get the Page Number from the physical address

PTE_ADDR

Get the address(or the frame number etc.) from the page table entry (PTE)

PTX

Get the Page Table Offset from the virtual address

ROUND

Round-up to the specified fractor.

ROUNDDOWN

Round-down to the specified fractor.

debug

Debug Formatted Print: with no new-line at the end. Will perform nothing in release profile.

debugln

Debug Formatted Print: with a new-line at the end. Will perform nothing in release profile.

pa2page

Get the page object of the phisical address.

page2kva

Get the kernel virtual address of the page object.

page2pa

Get the physical address of the page object.

page2ppn

Get the page number through the page object.

print

Formatted Print: with no new-line at the end.

println

Formatted Print: with a new-line at the end.

va2pa

Get the physical address of the given virtual address. This macro will look up the page table to do the transmition.

Statics

BUDDY_ALLOCATOR 🔒

Global Buddy Allocator. Can only alloc up to 32 KB at once.