rusty_mos::arch
Module machine

Machine related implementations.

Constants

KSEG1 🔒

The KSEG1 to visit the devices.

MALTA_FPGA_BASE 🔒

QEMU MMIO address definitions for FPGA (Used for halt).

MALTA_PCIIO_BASE 🔒

QEMU MMIO address definitions for PCIIO (Used for text I/O).

MALTA_SERIAL_BASE 🔒

16550 Serial UART device definitions (Base).

MALTA_SERIAL_DATA 🔒

16550 Serial UART device definitions (Data: Offset: 0x0).

MALTA_SERIAL_DATA_READY 🔒

Serial data is ready.

MALTA_SERIAL_LSR 🔒

16550 Serial UART device definitions (LSR: Offset: 0x5).

MALTA_SERIAL_THR_EMPTY 🔒

Serial the is empty.

Functions

halt

Actually trigger the reboot of the board. But in QEMU, we choose to ‘no-reboot’. So the simulation will exit if halt is called.

print_charc

Put the character (whose size is 1 byte) into teh serial. Busy-wait if the serial thr is not empty. The '\r' will be added before a '\n'.

scan_charc

Get a character (whose size is 1 byte) from the serial. Zero will be returned if the data is not ready for reading.