Module rusty_mos::arch_mipsel::syscall_impl

Syscall implementations.

Structs

CLikeStr 🔒

A wrapper for the *const u8.

Constants

SYSCALL_TABLE

Syscall function table. Indexed with the syscall number.

Functions

do_syscall

Get the syscall number and all the five arguments. Invoke the syscall.

sys_bind_shared_pool 🔒

SYSNO: 19, bind a existed memory pool.

sys_cgetc 🔒

SYSNO: 15, read a char from the console via scan_charc.

sys_create_shared_pool 🔒

SYSNO: 18, create a shared memory pool.

sys_env_destroy 🔒

SYSNO: 4, destory a specified env with the id of envid.

sys_exofork 🔒

SYSNO: 9, fork a new env which will be the child of the current env.

sys_getenvid 🔒

SYSNO: 2, get the EnvData::id or the pid. (a.k.a in Linux) of the current env.

sys_ipc_recv 🔒

SYSNO: 14, wait for a ipc-message.

sys_ipc_try_send 🔒

SYSNO: 13, Try to send a ipc-message to the target env.

sys_lock 🔒

SYSNO: 20, lock a specified memory pool.

sys_mem_alloc 🔒

SYSNO: 6, alloc a page and map it to va.

sys_mem_map 🔒

SYSNO: 7, map a page from one env (src_id) to another (dst_id).

sys_mem_unmap 🔒

SYSNO: 8, cancel the map of the specified virtual address for the specified env.

sys_panic 🔒

SYSNO: 12, trigger the kernel’s panic with the given C-Style string.

sys_print_cons 🔒

SYSNO: 1, show one num-lengthed string on the console.

sys_putchar 🔒

SYSNO: 0, just show a character on the console.

sys_read_dev 🔒

SYSNO: 17, read data from pa into va with the length of len.

sys_set_env_status 🔒

SYSNO: 10, set the run status of the specified env.

sys_set_tlb_mod_entry 🔒

SYSNO: 5, set the specified env’s user_tlb_mod_entry.

sys_set_trapframe 🔒

SYSNO: 11, set the trapframe of the specified env.

sys_unlock 🔒

SYSNO: 21, unlock a memory pool.

sys_write_dev 🔒

SYSNO: 16, write data at va into pa with the length of len.

sys_yield 🔒

SYSNO: 3, give out the CPU, re-schedule. NO-RETURN

Type Aliases

SyscallFn 🔒

The real syscall function type.

SyscallRawPtr 🔒

Just a type used in the SYSCALL_TABLE. A holder.