Function rusty_mos::arch_mipsel::syscall_impl::sys_set_trapframe

fn sys_set_trapframe(envid: u32, trapframe: *mut TrapFrame) -> u32

SYSNO: 11, set the trapframe of the specified env.

The env shall be the current env or the child of the current env.

The raw pointer trapframe shall be in the [UTEMP, UTOP) range.

Failure

This syscall will return a negated-KError::Invalid if the trapframe is not out of the required range.

This syscall will return a negated-KError returned by envid2env.