Function rusty_mos::arch_mipsel::syscall_impl::sys_write_dev

fn sys_write_dev(va: u32, pa: u32, len: u32) -> u32

SYSNO: 16, write data at va into pa with the length of len.

The len can only be in 1, 2 or 4.

All the valid devices and their physical address ranges are as follows:

device	start address	length

console	0x180003f8	0x20

IDE disk	0x180001f0	0x8

See Also: sys_read_dev

Failure

This syscall will return a negated-KError::Invalid if the va is out of the [[UTEMP, UTOP]) range, or the pa is invalid or the len is invalid.