Module rusty_mos::kdef

Modules

bitops

cp0reg

Define all the status bits for the cp0 reg.

elf

Define the Elf32 struct and the basic contants for use.

env

Define the basic structure used in the env module. As well as their basic constructor use by the static initialization.

error

Define the Error Code in mos.

mmu

Definitions about the memory / page and conventions. Include some helper functions to do the conversion.

pmap

Definitions of the conversion macros for the page.

syscall

The syscall number enum definition.

types

Definition of some misc macros about type and offset.