Struct rusty_mos::kdef::elf::Elf32Ehdr

#[repr(C)]
pub struct Elf32Ehdr {
    pub entry: u32,
    /* private fields */
}

The ELF32 file header structure. The members are defined in the same order as in the actual elf file.

Fields
entry: u32

The virtual address for the entry point.

Implementations

impl Elf32Ehdr

pub fn from(binary: *const u8, size: usize) -> *const Self

Build the Elf32Ehdr object from the binary.

Return

The raw-pointer of the binary start if the size and the magic number is valid. Otherwise, a null point will be returned.

pub fn foreach(&self, apply: impl Fn(u32))

Walk all the program header entry, use the function apply passed.