Type Alias rusty_mos::kdef::env::EnvNode

pub type EnvNode = LinkNode<*mut EnvData>;

Env link node.

Aliased Type

struct EnvNode {
    pub next: *mut LinkNode<*mut EnvData>,
    pub prev: *mut *mut LinkNode<*mut EnvData>,
    pub data: *mut EnvData,
}

Fields
next: *mut LinkNode<*mut EnvData>

Pointing the next node. If this is the last node, the field will be null.

prev: *mut *mut LinkNode<*mut EnvData>

Pointing the previous node’s next field. If this is the first node, the field will point to the head’s head field.

data: *mut EnvData

The data stored in the link list, with the type T.

Implementations

impl EnvNode

pub const fn const_construct() -> Self

Used for the static construction. All members are filled with zero.