Module rusty_mos::kern::env

Core methods for the env module.

Structs

EnvsWrapper

Wrapper to make it aligned to a page.

Statics

ASID_BITMAP

Bitmap for the asid allocatoin.

BASE_PGDIR

The global pgdir.

CUR_ENV

The current env.

ENVS

The envs array used in the kernel space. The element in it has the link field to make it a link node.

ENVS_DATA

The envs array in kernel, mapped to the UENVS and used by the user program.

ENV_FREE_LIST

Free env list.

ENV_SCHE_LIST

Runnable env list.

PRE_ENV_RUN

Run before the env_run for tests only

Functions

env_alloc⚠

Alloc an env and setup its vm and PCB.

env_create⚠

Create an env and load the icode, set the priority. For tests mainly.

env_destory⚠

Destory an env and free it. Re-schedule will be performed.

env_free⚠

Free an env, and remove all its pages. The TLB will be flushed after the deletion of pages.

env_init

Init the env environment. Put the envs into the free list, and map the PAGES and ENVS to base_pgdir’s UPAGES and UENVS accordingly.

env_pop_tf⚠

Recover from exception, load the specified TrapFrame.

env_run⚠

Run the env. Save the CUR_ENV’s trapframe if CUR_ENV exists.

env_setup_vm⚠

Setup the virtual memory of the new-born env.

envid2env⚠

Get the env’s PCB by its id. If checkperm is set, the method will check whether the env just is CUR_ENV or the child of CUR_ENV. If not, KError::BadEnv will be returned with a Err wrapper.

load_icode⚠

Load the icode for the e.