rusty_mos::kernel::syscall
Enum SyscallNo

#[repr(u8)]
pub enum SyscallNo {

    SysPutchar = 0,
    SysPrintCons = 1,
    SysGetenvid = 2,
    SysYield = 3,
    SysEnvDestroy = 4,
    SysSetTlbModEntry = 5,
    SysMemAlloc = 6,
    SysMemMap = 7,
    SysMemUnmap = 8,
    SysExofork = 9,
    SysSetEnvStatus = 10,
    SysSetTrapframe = 11,
    SysPanic = 12,
    SysIpcTrySend = 13,
    SysIpcRecv = 14,
    SysCgetc = 15,
    SysWriteDev = 16,
    SysReadDev = 17,
    SysCreatePool = 18,
    SysBindPool = 19,
    SysLock = 20,
    SysUnlock = 21,

}

Syscall identifier. The number is ordered by enum automatically.

Variants

SysPutchar = 0

Print a char into the console.

SysPrintCons = 1

Print a straight string ends with zero into the console.

SysGetenvid = 2

Get the id of the current env.

SysYield = 3

Give out the CPU time and re-schedule.

SysEnvDestroy = 4

Destory a env by its id and kill it.

SysSetTlbModEntry = 5

Register the user-space TLB mod handler for the specified env.

SysMemAlloc = 6

Allocate memory and map it.

SysMemMap = 7

Map the virtual address to a specified physical page.

SysMemUnmap = 8

Unmap the address and the page.

SysExofork = 9

Allocate a new env and make it child of the current env.

SysSetEnvStatus = 10

Set the env status and move it between the lists.

SysSetTrapframe = 11

Set the trapframe to he specified env.

SysPanic = 12

Do kernel panic.

SysIpcTrySend = 13

Try to send an ipc data to a env.

SysIpcRecv = 14

Receive a ipc data or wait.

SysCgetc = 15

Scan a char from the console

SysWriteDev = 16

Write to a dev.

SysReadDev = 17

Read from a dev.

SysCreatePool = 18

Create a memory pool.

SysBindPool = 19

Bind a memory pool.

SysLock = 20

Try to lock a memory pool.

SysUnlock = 21

Unlock a memory pool.