rusty_mos::kernel::syscall_impl
Function sys_exofork

fn sys_exofork() -> u32

SYSNO: 9, fork a new env which will be the child of the current env.

For the child, the priority will be same with its parent and the trap frame also. The ruturn value will be set to zero to mark the child.

Failure

This syscall will return a negated-KError returned by env_alloc.