rusty_mos::kernel::syscall_impl
Function sys_ipc_recv

fn sys_ipc_recv(dst_va: u32) -> u32

SYSNO: 14, wait for a ipc-message.

The message will be a value together with a page if the dst_va is not zero.

The current env will be blocked.

Failure

This syscall will return a negated-KError::Invalid if the dst_va is not zero and is not in the [[UTEMP, UTOP]) range.

This syscall will not return if everything goes smooth.