rusty_mos::kernel::syscall_impl
Function sys_mem_map

fn sys_mem_map(
    src_id: u32,
    src_va: u32,
    dst_id: u32,
    dst_va: u32,
    perm: u32,
) -> u32

SYSNO: 7, map a page from one env (src_id) to another (dst_id).

Get the page with the virtual address (src_va) in the env (src_id), and map it into the virtual address (dst_va) in the env (dst_id).

Both the source env and the destination env shall be the current env or the child of the current env.

See Also: sys_mem_unmap

Failure

This syscall will return a negated-KError returned by envid2env, page_lookup and page_insert.

This syscall will return a negated-KError::Invalid if any of the virtual address is above the UTOP or below the UTEMP.