rusty_mos::kernel::syscall_impl
Function sys_mem_unmap

fn sys_mem_unmap(envid: u32, va: u32) -> u32

SYSNO: 8, cancel the map of the specified virtual address for the specified env.

The env shall be the current env or the child of the current env.

Failure

This syscall will return a negated-KError returned by envid2env and page_remove.

This syscall will return a negated-KError::Invalid if the virtual address is above the UTOP or below the UTEMP.