Struct rusty_mos::memory::buddy_allocator::BuddyInner

struct BuddyInner<const CCOUNT: usize> {
    free_list: [PageList; CCOUNT],
    page_start: *mut PageNode,
}

The buddy system core data structure.

Fields
free_list: [PageList; CCOUNT]

Free page lists. Only the first page will be in the list.

The i-th category holds (2^i) pages each.

page_start: *mut PageNode

The start page address used in the system.

Implementations

impl<const CCOUNT: usize> BuddyInner<CCOUNT>

const fn new() -> Self

Create a brand-new buddy contents.

fn init(&mut self, page_start: *mut PageNode, size: usize)

Initialize the free list. The pages will be inserted into the topest category and the remained into one-level-lower category. As to the category zero.

Only after this function invoked, the system can be used.

impl<const CCOUNT: usize> BuddyInner<CCOUNT>

fn alloc(&mut self, layout: Layout) -> *mut u8

Buddy Alloc.

This method will search the specified category’s free list first. If found, the page will be returned.

Otherwise, this method will search the higher category, until a free item was found. In this situation, the pages will be splitted and inserted into the lower category’s free list one by one.

If no pages can be found, a null pointer will be returned.

fn dealloc(&mut self, ptr: *mut u8, layout: Layout)

Buddy Dealloc.

The page will be inserted into its category’s free list if no buddy is found.

Otherwise, the buddy will be removed from the free list. They will be merged into a larger item and be inserted into one-level-higher category.

The process above will be performed recursively unless reached the toppest category.