rusty_mos::memory
Module pmap

The page-memory manager model of MOS. Core functions are provided here.

Structs

PageData

The only page data needed to maintain.

Constants

PAGE_SIZE 🔒

The page size is 4 KB.

Statics

CUR_PGDIR

Current env’s page directary address.

KERN_HEAP

The kernel heap start. Used by the buddy system.

NPAGE

The the count of all the pages available.

PAGES

The kernel array for all pages.

PAGE_FREE_LIST

The list of free pages.

Functions

alloc 🔒

Alloc needed memorys in kernel mode. Only use it before the page-manager is ready to use.

mips_detect_memory

Calculate the NPAGE and validate it.

mips_vm_init

Alloc memories needed for the PAGES and reserve 512 * 4KB spaces for the buddy system to alloc.

page_alloc

Alloc a page. Return the page’s address or an error if no free pages available.

page_decref

Decrease the page’s pp_ref. If all reference is removed, the page will be freed.

page_free

Free a page. The pp_ref shall be zero before freeing it.

page_init

Init all the pages. Mark the pages below the freemem as used.

page_insert

Map the physical page to the virtual address va. The permission bits will be set to perm | PTE_C_CACHEABLE | PTE_V.

page_lookup

Look up the Page that virtual address va map to. Return the page and the page table entry together if the page is found and is valid.

page_remove

Unmap the physical page at virtual address va.

pgdir_walk

Walk the current page table to find the virtual address va’s page table entry (Pte).

Type Aliases

PageList

Page list, for the ‘free_list’. See Also: PAGE_FREE_LIST.

PageNode

Node in the page list.

Pde

The page directory entry type alias.

Pte

The page table entry type alias.