rusty_mos::memory::pmap
Type Alias PageList

pub type PageList = LinkList<PageData>;

Page list, for the ‘free_list’. See Also: PAGE_FREE_LIST.

Aliased Type

struct PageList {
    pub head: *mut LinkNode<PageData>,
}

Fields
head: *mut LinkNode<PageData>

Pointing to the first node of this link list. The list is empty if and only if the head is null.