Module rusty_mos::memory::regions

Definitions about the memory / page and conventions. Include some helper functions to do the conversion.

The virtual address (32-bit) structure is as follows:

| 31                 22 | 21             12 | 11           0 |
| Page Directory Offset | Page Table Offset | In-Page Offset |

The page table entry (PTE, 32-bit) structure is as follows:

| 31                 12 | 11               6 | 5                0 |
| Physical Frame number | Flags for hardware | Flags for software |

Constants

KERNBASE

Kernel memory starts from here

KSEG1

KSEG1 Segment

KSTACKTOP

Kernel stack end at here (the end in the linking script).

NASID

The maximum count of all the asid

PAGE_SIZE

The physical page size (in bytes).

PDMAP

Bytes mapped by a page directory entry.

PDSHIFT

Shifted the Page Table Offset and In-Page Offset out to get the Page Directory Offset

PGSHIFT

Shifted the In-Page Offset out to get the Page Table Offset

PTE_C_CACHEABLE

Cache Coherency Attributes bit. If set, this entry is cache-able.

PTE_D

Dirty bit, but really a write-enable bit. 1 to allow writes, 0 and any store using this translation will cause a tlb mod exception (TLB Mod).

PTE_G

Global bit. When this bit in a TLB entry is set, that TLB entry will match solely on the VPN field, regardless of whether the TLB entry’s ASID field matches the value in EntryHi.

PTE_HARDFLAG_SHIFT 🔒

Page table / directory entry flag shift

PTE_V

Valid bit. If 0 any address matching this entry will cause a tlb miss exception (TLBL/TLBS).

PTMAP 🔒

Bytes mapped by a page table entry (Actually is the PAGE_SIZE).

UCOW

Reserved for COW (start address).

UENVS

The kernel array ENVS will be mapped here.

ULIM

The high-limits of user’s memory

UPAGES

The kernel array PAGES will be mapped here.

USTACKTOP

Normal user stack top.

UTEMP

Reserved for temporary usage (start address).

UTEXT

User test segment start.

UTOP

The uer’s space higher boundary.

UVPT

User’s page tables are stored here (for a PDMAP size).

UXSTACKTOP

The exception stack top for the user. See also: UTOP.