rusty_mos::memory
Module shared_pool

Support the shared momeory pool. We can share more than one pages between envs.

Structs

MemoryPool

Memory pool manager. Support every operation that is needed to perform.

MemoryPoolEntry 🔒

Memory Pool Entry. Contains the pages it shared and the envs attatched to it. The entry also contains a lock field.

Statics

MEMORY_POOL

Global shared memory pool.

POOL_I 🔒

Used to spawn the pool id. Increase one-by-one when doing mkpoolid.

Functions

mkpoolid 🔒

Spawn the unique id of a new pool.