Struct rusty_mos::memory::shared_pool::MemoryPoolEntry

struct MemoryPoolEntry {
    pages: Vec<usize>,
    envs: Vec<usize>,
    reference: usize,
    write_lock: AtomicUsize,
}

Memory Pool Entry. Contains the pages it shared and the envs attatched to it. The entry also contains a lock field.

Fields
pages: Vec<usize>

List of page indexes shared by this pool.

envs: Vec<usize>

List of envs bind to this pool.

reference: usize

The reference count of this pool. Destroy the pool is the reference goes zero.

write_lock: AtomicUsize

Write lock, store the env-id which locked it. Or zero means unlocked.

Implementations

impl MemoryPoolEntry

pub const fn new() -> Self

Create a new pool.

fn deref(&mut self, envid: usize) -> Result<bool, KError>

Decrease the reference of the pool. If the reference is minused to zero, an Ok(true) will be returned.

pub fn lock(&mut self, envid: usize) -> bool

Try to lock the pool. If locked successfully, the return value is true.

pub fn unlock(&mut self, envid: usize) -> Result<(), KError>

Unlock the pool. Only the env who locked it can unlock it.

Return

If unlocked successfully, an Ok(()) is returned.

Otherwise, a KError with a Err wrapper is returned:

Err(KError::NoLock): The pool was not locked.

Err(KError::LockByOthers): The pool was locked by another env.