rusty_mos
Module utils

Kernel independent utils like link list or bit map.

Modules

array_based_list

Safe linked list managed by arrays.

bitmap

Provide a bitmap and related opeartions.

bitops

Bit operations.

io

Just read/write from the device address.

linked_list

Link List implemented with Rust, which is similar to the kernel queue of the mos / Linux

sync_ref_cell

RefCell wrapper. It implements Sync for global use.