Struct rusty_mos::utils::array_based_list::ArrayLinkedList

#[repr(C)]
pub struct ArrayLinkedList<const LEN: usize> {
    array: [ArrayLinkNode; LEN],
    head: Option<usize>,
    tail: Option<usize>,
}

The list itself. Maintain a head and tail and the node array.

Generic

The const LEN means the node count. It should be specified at the compile time.

Invariant

Both head and tail shall be None together or not be None together.

See Also: ArrayLinkNode

Fields
array: [ArrayLinkNode; LEN]

The array storing the linking field of each nodes.

head: Option<usize>

The index of the first node. None means an empty node.

tail: Option<usize>

The index of the last node. None means an empty node.

Implementations

impl<const LEN: usize> ArrayLinkedList<LEN>

pub const fn new() -> Self

Create a new link list and initialize the array.

pub fn empty(&self) -> bool

Judge whether the list is empty.

pub fn insert_head(&mut self, item: usize)

Insert the index item node into the head.

pub fn peek_head(&self) -> Option<usize>

Peek the first node.

pub fn pop_head(&mut self) -> Option<usize>

Get the first node. And remove it from the list if the list is not empty.

pub fn insert_tail(&mut self, item: usize)

Insert the index item node into the tail.

pub fn remove(&mut self, item: usize)

Remove the index item node from the list. This method requires that the node was in the list.

pub fn contains(&self, item: usize) -> bool

Judge whether the node is in the list.