rusty_mos::utils::bitmap
Struct Bitmap

pub struct Bitmap<const COUNT: usize> {
    bitmap: [u8; COUNT],
    used: usize,
    pointer: usize,
}

The bitmap itself. Contains the map and assistant field.

The used maintains the count of bit used to speed up the empty or full judgement.

The algorithm of searching the next empty bit is as follows:

We maintain a pointer, the search will from the pointer one-by-one.

Generic

The const CCOUNT marks the count of bit maintains. The bitmap can hold (CCOUNT * 8) bits.

Fields
bitmap: [u8; COUNT]

The bitmap itself.

used: usize

Used bits count.

pointer: usize

Search pointer.

Implementations

impl<const COUNT: usize> Bitmap<COUNT>

pub const fn new() -> Self

Create a new bitmap and initialize the field as zero.

pub fn empty(&self) -> bool

Judge whether the bitmap is empty.

pub fn peek(&self, index: usize) -> bool

To see whether the index is not used.

pub fn alloc(&mut self) -> Option<usize>

Alloc a new bit if available. A None will be returned if no bit rest.

pub fn free(&mut self, index: usize)

Free the specified bit. Double free is not allowed.