rusty_mos::utils::linked_list
Struct LinkList

pub struct LinkList<T: Copy> {
    pub head: *mut LinkNode<T>,
}

The head struct of the LinkList

See also: LinkNode

Generics

The type T indicates the data stored in the link list.

Fields
head: *mut LinkNode<T>

Pointing to the first node of this link list. The list is empty if and only if the head is null.

Implementations

impl<T: Copy> LinkList<T>

pub const fn new() -> LinkList<T>

Create an empty link list with its head null.

pub fn empty(&self) -> bool

Judge whether this list is empty.

pub fn insert_head(&mut self, item: *mut LinkNode<T>)

Insert a node to the head of the list

Safety

The parameter item SHALL be mutably-visitable!

pub fn pop_head(&mut self) -> Option<*mut LinkNode<T>>

Get the first node of this list and removce it

The return value will be None is the list is empty.

Safety

All things in the list SHALL be valid!

pub fn remove(item: *mut LinkNode<T>)

Remove a specified node from the list contains this node.

Safety

The parameter item SHALL be mutably-visitable and SHALL be in an valid link list!