rusty_mos::utils::linked_list
Struct LinkNode

pub struct LinkNode<T: Copy> {
    pub next: *mut LinkNode<T>,
    pub prev: *mut *mut LinkNode<T>,
    pub data: T,
}

The node struct of the LinkList

See also: LinkList

Generics

The type T indicates the data stored in the link list.

Fields
next: *mut LinkNode<T>

Pointing the next node. If this is the last node, the field will be null.

prev: *mut *mut LinkNode<T>

Pointing the previous nodeโs next field. If this is the first node, the field will point to the headโs head field.

data: T

The data stored in the link list, with the type T.

Implementations

impl<T: Copy> LinkNode<T>

pub const fn new(data: T) -> LinkNode<T>

Create an empty link list node with its linking-field all null.