Struct rusty_mos::utils::linked_list::TailLinkList

pub struct TailLinkList<T: Copy> {
    pub head: *mut LinkNode<T>,
    pub tail: *mut *mut LinkNode<T>,
}

The LinkList with its tail recorded

See also: LinkNode

Attention

Before use the TailLinkList object, the method enable SHALL be called first.

Generics

The type T indicates the data stored in the link list.

Fields
head: *mut LinkNode<T>tail: *mut *mut LinkNode<T>
Implementations

impl<T: Copy> TailLinkList<T>

pub const fn new() -> TailLinkList<T>

Create an empty TailLinkList object. Not useable yet.

pub fn empty(&self) -> bool

Judge whether the list is empty.

pub fn enable(&mut self)

Make the list useable after the construction.

pub unsafe fn insert_head(&mut self, item: *mut LinkNode<T>)

Insert the item into the head of the list.

Safety

The item passed SHALL NOT be null.

pub unsafe fn insert_tail(&mut self, item: *mut LinkNode<T>)

Insert the item into the tail of the list.

Safety

The item passed SHALL NOT be null.

pub unsafe fn pop_head(&mut self) -> Option<*mut LinkNode<T>>

Get the head and remove it if the list is not empty.

Return

None – The list is empty. Some(item) – Otherwise

Safety

The list SHALL be valid.

pub unsafe fn remove(&mut self, item: *mut LinkNode<T>)

Remove a specified node from the list contains this node.

Safety

The parameter item SHALL be mutably-visitable and SHALL be in an valid link list!