Crate rusty_mos

Modules

arch_mipsel

consts

library

memory

process

utils

Macros

ARRAY_PTR

Get the array[i] via the raw pointer.

ENVX

Spawn the index of the given envid.

GEN_MASK

Gen a mask from the $h-bit (MSB) to $l-bit (LSB) with 1.

KADDR

Get the virtual address (in kernel segment) from the physical address

PADDR

Get the physical address of the virtual address in kernel segment

PDX

Get the Page Directory Offset from the virtual address

PPN

Get the Page Number from the physical address

PTE_ADDR

Get the address(or the frame number etc.) from the page table entry (PTE)

PTX

Get the Page Table Offset from the virtual address

ROUND

Round-up to the specified fractor.

ROUNDDOWN

Round-down to the specified fractor.

debug

debugln

pa2page

Get the page object of the phisical address.

page2kva

Get the kernel virtual address of the page object.

page2pa

Get the physical address of the page object.

page2ppn

Get the page number through the page object.

print

println

va2pa

Get the physical address of the given virtual address. This macro will look up the page table to do the transmition.