Module rusty_mos::process::elf_loader

Structs

Elf32Ehdr

The ELF32 file header structure. The members are defined in the same order as in the actual elf file.

Elf32Phdr

The program segment header structure. The members are defined in the same order as in the actual elf file.

Constants

PF_R

Mark the segment as readable.

PF_W

Mark the segment as writable.

PF_X

Mark the segment as executable.

PT_LOAD

Mark the segment as loadble and load-needed.

Functions

elf_load_seg

Load an elf-format binary file in memory. This method will map all sections to correct virtual address.

Type Aliases

ElfMapperFn

The type used in the walker of the header.