Module rusty_mos::process::envs

Core methods for the env module.

Structs

EnvData

The PCB struct. Compatible with the C-Like memory structure and the MOS.

EnvsWrapper

Wrapper to make it aligned to a page.

IpcData

The IPC data collected together. Compatible with the C-Like memory structure.

Enums

EnvStatus

The env status enum. Compatible with the C-Like memory structure.

Constants

LOG2NENV

The log of NENV.

NENV

The count of the envs.

Statics

ASID_BMAP

The envs array used in the kernel space. The element in it has the link field to make it a link node. Bitmap for the asid allocatoin.

BASE_PGDIR

The global pgdir.

CUR_ENV_IDX

The current env.

ENVS_DATA

The envs array in kernel, mapped to the UENVS and used by the user program.

ENV_FREE_LIST

Free env list.

ENV_I 🔒

The static variable used by mkenvid;

ENV_SCHE_LIST

Runnable env list.

Functions

asid_alloc 🔒

Alloc a new asid. Return an error if there is no asid remained.

asid_free 🔒

Free an asid.

env_alloc

Alloc an env and setup its vm and PCB.

env_create

Create an env and load the icode, set the priority. For tests mainly.

env_destory

Destory an env and free it. Re-schedule will be performed.

env_free

Free an env, and remove all its pages. The TLB will be flushed after the deletion of pages.

env_init

Init the env environment. Put the envs into the free list, and map the PAGES and ENVS to base_pgdir’s UPAGES and UENVS accordingly.

env_pop_tf⚠

Recover from exception, load the specified TrapFrame.

env_run

Run the env. Save the current env’s trapframe if it exists.

env_setup_vm

Setup the virtual memory of the new-born env.

envid2env

Get the env’s PCB by its id. If checkperm is set, the method will check whether the env just is the current env or the child of it. If not, KError::BadEnv will be returned with a Err wrapper.

load_icode 🔒

Load the icode for the e.

map_segment 🔒

Map the [va, va + size) in virtual address space to the [pa, pa + size) in physical address space if pgdir. This method will validate the entry perm.

mkenvid 🔒

Get the envid with the env node.